+++ src/LLAPI/Renderer/Renderer/MeshElement.cpp
@@ -11,8 +11,8 @@ namespace oakvr
 		{
 			m_vertexFormat.push_back(vertexFormat);
 			m_vertexStride = vertexFormat.size;
-			m_indexCount = m_indexData.Size() / m_indexStride;
-			m_vertexCount = m_vertexData.Size() / m_vertexStride;
+			m_indexCount = static_cast<uint32_t>(m_indexData.Size() / m_indexStride);
+			m_vertexCount = static_cast<uint32_t>(m_vertexData.Size() / m_vertexStride);
 		}
 
 		MeshElement::MeshElement(const std::vector<VertexElementDescriptor> &vertexFormat, const oakvr::core::MemoryBuffer &vb
@@ -25,8 +25,8 @@ namespace oakvr
 			{
 				m_vertexStride += e.size;
 			}
-			m_indexCount = m_indexData.Size() / m_indexStride;
-			m_vertexCount = m_vertexData.Size() / m_vertexStride;
+			m_indexCount = static_cast<uint32_t>(m_indexData.Size() / m_indexStride);
+			m_vertexCount = static_cast<uint32_t>(m_vertexData.Size() / m_vertexStride);
 		}
 
 		MeshElement::~MeshElement()
